args: created to store cli command parsing
I'm using clap because it is simple and easy to use and makes the chore
of parsing cli arguments take only a few seconds to maintain and create.

The clap library currently works by defining arguments in a struct using
derive macros. This struct is stored here, in a separate file.

Signed-off-by: Dylan Turner <dylantdmt@gmail.com>