README: first set of extensive documentation
I hope to make this project fully understandable even by those with
little Rust experience or little programming experience, so that
contributors feel welcome.

As such, I've tried to lay out a decent amount of basic info on how the
project works as well as user-level documentation like how the commands
work and how to install the program.

This will need to be maintained and updated as the project grows, but I
believe this is a good starting point.

Signed-off-by: Dylan Turner <dylantdmt@gmail.com>